Move renameAndSort inside serialize
